Rename model files to have the _model suffix
