+++ apeiros/models/location_model.py
@@ -4,7 +4,7 @@ from sqlalchemy import ForeignKey, BLOB
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 
-from .base import Base
+from .base_model import Base
 from .enums import LocationTag
