+++ apeiros/models/medal_model.py
@@ -1,7 +1,7 @@
 from typing import Optional
 
 from sqlalchemy.orm import Mapped, mapped_column
-from .base import Base
+from .base_model import Base
 
 
 class Medal(Base):
