+++ apeiros/models/nomster_model.py
@@ -3,7 +3,7 @@ from typing import Optional
 from sqlalchemy import ForeignKey, BLOB
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
-from .base import Base
+from .base_model import Base
 from .enums import NomsterType, NomsterClass
