+++ apeiros/models/player_model.py
@@ -3,10 +3,10 @@ from typing import Optional, List, Dict, Any
 from sqlalchemy import Table, Column, ForeignKey, BLOB
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
-from .base import Base
-from .medal import Medal
-from .item import Item
-from .nomster import Nomster
+from .base_model import Base
+from .medal_model import Medal
+from .item_model import Item
+from .nomster_model import Nomster
 
 
 _player_medal_association_table = Table(
